Add prompt injection feature
With M1+NUMPAD_8 you can now edit an additional prompt. This prompt is
added as Java comment to the start of the completion prefix. It might
not work in all cases, but it might help to improve the Tabnine results
by specifying libraries that are available or the Java source level that
should be used.